+++ cmd/grainfs/serve_nbd.go
@@ -9,8 +9,11 @@ import (
 	"github.com/gritive/GrainFS/internal/volume"
 )
 
-func startNBDServer(mgr *volume.Manager, volName string, port int) (*nbd.Server, error) {
+func startNBDServer(mgr *volume.Manager, volName string, port int, ri nbd.ReadIndexer) (*nbd.Server, error) {
 	srv := nbd.NewServer(mgr, volName)
+	if ri != nil {
+		srv.SetReadIndexer(ri)
+	}
 	go func() {
 		addr := fmt.Sprintf(":%d", port)
 		if err := srv.ListenAndServe(addr); err != nil {